29 apr da pc, creata interfaccia cli con funzionalità principali (da completare)

+++ src/main/java/com/example/spring/spring/model/persona/Cliente.java
@@ -14,7 +14,7 @@ public class Cliente extends Persona {
 
     }
 
-    public Cliente(String nome, String cognome, int eta, int telefono, String email) {
+    public Cliente(String nome, String cognome, int eta, long telefono, String email) {
         super(nome, cognome, eta, telefono);
         this.email = email;
     }
